+++ src/App/exports/Remove-AzContainerAppManagedEnvCert.ps1
@@ -47,7 +47,7 @@ INPUTOBJECT <IAppIdentity>: Identity Parameter
   [StorageName <String>]: Name of the storage.
   [SubscriptionId <String>]: The ID of the target subscription.
 .Link
-https://docs.microsoft.com/powershell/module/az.app/remove-azcontainerappmanagedenvcert
+https://learn.microsoft.com/powershell/module/az.app/remove-azcontainerappmanagedenvcert
 #>
 function Remove-AzContainerAppManagedEnvCert {
 [OutputType([System.Boolean])]