+++ src/Compute/Compute.Autorest/custom/Set-AzVmssVMRunCommand_ScriptLocalPath.ps1
@@ -16,7 +16,7 @@
 
 function Set-AzVmssVMRunCommand_ScriptLocalPath {
     [OutputType([Microsoft.Azure.PowerShell.Cmdlets.Compute.Models.Api20210701.IVirtualMachineRunCommand])]
-    [CmdletBinding(DefaultParameterSetName='UpdateExpanded', PositionalBinding=$false, SupportsShouldProcess, ConfirmImpact='Medium')]
+    [CmdletBinding(PositionalBinding=$false, SupportsShouldProcess, ConfirmImpact='Medium')]
     param(
         [Parameter(Mandatory)]
         [Microsoft.Azure.PowerShell.Cmdlets.Compute.Category('Path')]
@@ -186,12 +186,12 @@ function Set-AzVmssVMRunCommand_ScriptLocalPath {
         if ($PSBoundParameters.ContainsKey("ScriptLocalPath"))
         {
             # Read Local File and add 
-            $script = Get-Content -Path $ScriptLocalPath
+            $script = (Get-Content -Path $ScriptLocalPath) -join ";"
             $PSBoundParameters.Add("SourceScript", $script)
             # If necessary, remove the -ParameterA parameter from the dictionary of bound parameters
             $null = $PSBoundParameters.Remove("ScriptLocalPath")
         }
-        Az.Compute\Set-AzVMRunCommand @PSBoundParameters
+        Az.Compute\Set-AzVmssVMRunCommand @PSBoundParameters
     }
 }