+++ src/DataProtection/Changelog.md
@@ -19,6 +19,9 @@
 -->
 ## Upcoming Release
 
+## Version 1.0.1
+* Fixed spacing issues in Set-AzDataProtectionMSIPermission.ps1
+
 ## Version 1.0.0
 * Fixed list parameter set for `Get-AzDataProtectionBackupVault`