+++ src/Network/Network/ChangeLog.md
@@ -20,6 +20,15 @@
 
 ## Upcoming Release
 
+## Version 5.1.2
+* Upgraded AutoMapper to Microsoft.Azure.PowerShell.AutoMapper 6.2.2 with fix [#18721]
+
+## Version 5.1.1
+* Fixed bug with New-AzCustomIpPrefix
+
+## Version 5.1.1
+* Fixed bug with New-AzCustomIpPrefix
+
 ## Version 5.1.0
 * Added possible value `LocalGateway` for parameter `GatewayType`
     - `New-AzVirtualNetworkGateway`