+++ src/RecoveryServices/RecoveryServices/ChangeLog.md
@@ -18,6 +18,11 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Fixed the pagination bug in `Get-AzRecoveryServicesAsrProtectableItem` for the V2ARCM scenario.
+* Fixed `IncludeDiskId` property for `New-ASRReplicationProtectedItem` cmdlet of H2A
+
+## Version 6.1.1
+* Upgraded AutoMapper to Microsoft.Azure.PowerShell.AutoMapper 6.2.2 with fix [#18721]
 
 ## Version 6.1.0
 * Added support for cross zonal restore for ZRS vaults for non-ZonePinned VM